requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-cors-headers==3.2.1
django-jsonfield==1.4.0
djangorestframework==3.11.0
idna==2.8
lxml==4.4.2
nltk==3.4.5
pytz==2019.3
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
textblob==0.15.3
urllib3==1.25.8
